+++ packages/hydrogen/src/components/index.ts
@@ -44,7 +44,6 @@ export {ProductDescription} from './ProductDescription';
 export {ProductTitle} from './ProductTitle';
 export {ProductPrice} from './ProductPrice';
 export {ProductMetafield} from './ProductMetafield';
-export {SelectedVariantPrice} from './SelectedVariantPrice';
 export {SelectedVariantUnitPrice} from './SelectedVariantUnitPrice';
 export type {
   BuyNowButtonProps,